refactor: Rename houseOverview directory to house
