+++ app/src/main/java/com/example/dtthouses/ui/house/viewModel/HouseViewModel.kt
@@ -1,7 +1,6 @@
-package com.example.dtthouses.ui.houseOverview.viewModel
+package com.example.dtthouses.ui.house.viewModel
 
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.ViewModel
 import com.example.dtthouses.data.model.House
 import com.example.dtthouses.utils.Resource
